+++ module/sheets/actor-sheet.mjs
@@ -1466,16 +1466,6 @@ async _onUseFavorite(event) {
   /* -------------------------------------------- */
 
   /**
-   * Handle toggling inspiration.
-   * @protected
-   */
-  _onToggleInspiration() {
-    this.actor.update({ "system.attributes.inspiration.status": !this.actor.system.attributes.inspiration.status });
-  }
-
-  /* -------------------------------------------- */
-
-  /**
    * Initialize item tooltips on an element.
    * @param {HTMLElement} element  The tooltipped element.
    * @protected