+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Removed unlicenced Dungeons and Dragons Artworks (#36)
 - Renamed _baseattributes_ to _attributes_ on actor data (#27)
 - Added dice icons to rollable items and optimized css in chatcard presentation 
+- Removed Homebrew Inspiration points, as this should be a module (#40)